# chore: improve eslint performance